tests/requirements.txt
behave==1.2.6
configobj==5.0.6
PyHamcrest==2.0.2
requests==2.23.0
